moved TokenRange out of Ast
