+++ src/parser/nodeIdMap/nodeIdMapUtils.ts
@@ -3,6 +3,7 @@
 
 import { Ast, ParserContext } from "..";
 import { CommonError, isNever, Option } from "../../common";
+import { TokenRange } from "../../lexer";
 import {
     AstNodeById,
     ChildIdsById,
@@ -560,7 +561,7 @@ export function expectAncestry(nodeIdMapCollection: Collection, rootId: number):
 export function xorNodeTokenRange(nodeIdMapCollection: Collection, xorNode: TXorNode): XorNodeTokenRange {
     switch (xorNode.kind) {
         case XorNodeKind.Ast: {
-            const tokenRange: Ast.TokenRange = xorNode.node.tokenRange;
+            const tokenRange: TokenRange = xorNode.node.tokenRange;
             return {
                 tokenIndexStart: tokenRange.tokenIndexStart,
                 tokenIndexEnd: tokenRange.tokenIndexEnd,
